pyproject.toml
[tool.poetry]
name = "luisciber"
version = "0.0.1"
description = "Luis Ciber's Workspace"
authors = ["Luis Ciber <luisciber640@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.0"
mkdocstrings = "^0.25.1"
mkdocstrings-python = "^1.10.5"
mkdocs-jupyter = "^0.24.8"
mkdocs-redirects = "^1.2.1"
mkdocs-minify-plugin = "^0.8.0"
mkdocs-rss-plugin = "^1.15.0"
mkdocs-git-committers-plugin-2 = "^2.3.0"
markdown-include = "^0.8.1"
markdown-callouts = "^0.4.0"
mkdocs-material = {extras = ["imaging"], version = "^9.5.29"}


[tool.poetry.group.dev.dependencies]
pillow = "^10.4.0"
typer = "^0.12.3"


[tool.poetry.group.jupyter.dependencies]
jupyter = "^1.0.0"
ipywidgets = "^8.1.3"
ipykernel = "^6.29.5"
tqdm = "^4.66.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
